--- finalproject/lib/Hotels_page.dart
@@ -1,1 +1,0 @@
-// TODO Implement this library.
