+++ finalproject/lib/views/dashboard/hotel/add_hotel.dart
@@ -6,10 +6,11 @@ import 'dart:io';
 import 'dart:typed_data';
 
 // import 'package:image_picker/image_picker.dart';
-import 'package:project4/repository/user_repo.dart';
+import 'package:project4/repository/hotel_repo.dart';
 
-import '../CustomPages/appbar.dart';
-import '../models/user_model.dart';
+import '../../../CustomPages/appbar.dart';
+import '../../../models/hotel_model.dart';
+
 
 //
 // selectImageFromGallery() async {
@@ -23,20 +24,16 @@ import '../models/user_model.dart';
 // }
 
 
-class userAdd extends StatefulWidget {
-  const userAdd({Key? key}) : super(key: key);
+class hotelAdd extends StatefulWidget {
+  const hotelAdd({Key? key}) : super(key: key);
 
   @override
-  State<userAdd> createState() => _userAdd();
+  State<hotelAdd> createState() => _hotelAdd();
 }
 
-class _userAdd extends State<userAdd> {
+class _hotelAdd extends State<hotelAdd> {
   @override
-  String? selectedPermission;
-  String? selectedGender;
 
-  final permissionList = ['admin', 'user'];
-  final genderList = ['male', 'female', 'others'];
 
   String selectedImagePath = 'jhgggj';
 
@@ -46,12 +43,13 @@ class _userAdd extends State<userAdd> {
   String error="";
   var nameCtr=TextEditingController();
 
-  var phoneCtr=TextEditingController();
-  var emailCtr=TextEditingController();
-  var passwordCtr=TextEditingController();
-  var genderCtr=TextEditingController();
-  var perCtr=TextEditingController();
-  var usernameCtr=TextEditingController();
+  var descnmaeCtr=TextEditingController();
+  var descriptionCtr=TextEditingController();
+  var priceCtr=TextEditingController();
+  var locationCtr=TextEditingController();
+  var roomsCtr=TextEditingController();
+ // var avatarCtr=TextEditingController();
+
 
 
 
@@ -148,12 +146,12 @@ class _userAdd extends State<userAdd> {
               padding: const EdgeInsets.all(20.0),
               child: TextFormField(
 
-                controller: usernameCtr,
+                controller: descnmaeCtr,
                 autovalidateMode: AutovalidateMode.onUserInteraction,
                 style: TextStyle(color: Colors.black),
                 decoration: InputDecoration(
-                  hintText: "User Name",
-                  labelText: "User Name",
+                  hintText: "Desdescription Name",
+                  labelText: "Desdescription Name",
                   labelStyle: TextStyle(fontWeight: FontWeight.bold),
                   counterText: "20",
                   border:
@@ -172,7 +170,7 @@ class _userAdd extends State<userAdd> {
                     return "value is null";}
                   if(val != null){
                     if(val.length <3){
-                      return "User Name must be more than 3 chars";
+                      return "Desdescription Name must be more than 3 chars";
                     }
 
                   }
@@ -185,12 +183,12 @@ class _userAdd extends State<userAdd> {
               padding: const EdgeInsets.all(20.0),
               child: TextFormField(
 
-                controller: emailCtr,
+                controller: descriptionCtr,
                 autovalidateMode: AutovalidateMode.onUserInteraction,
                 style: TextStyle(color: Colors.black),
                 decoration: InputDecoration(
-                  hintText: "email",
-                  labelText: "email",
+                  hintText: "description",
+                  labelText: "desdescription",
                   labelStyle: TextStyle(fontWeight: FontWeight.bold),
                   counterText: "20",
                   border:
@@ -218,54 +216,17 @@ class _userAdd extends State<userAdd> {
 
               ),
             ),
+
             Padding(
               padding: const EdgeInsets.all(20.0),
               child: TextFormField(
 
-                controller: phoneCtr,
-                keyboardType: TextInputType.number,
+                controller: priceCtr,
                 autovalidateMode: AutovalidateMode.onUserInteraction,
                 style: TextStyle(color: Colors.black),
                 decoration: InputDecoration(
-                  hintText: "phone",
-                  labelText: "phone",
-                  labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                  counterText: "20",
-                  border:
-                  OutlineInputBorder(borderSide:BorderSide(color: Colors.amber),
-                      borderRadius: BorderRadius.circular(20)),
-                  errorBorder:  OutlineInputBorder(borderSide:BorderSide(color: Colors.red),
-                      borderRadius: BorderRadius.circular(20)),
-
-
-
-                ),
-
-
-                validator: (val){
-                  if(val == ""){
-                    return "value is null";}
-                  if(val != null){
-                    if(val.length <3){
-                      return "phone must be more than 3 chars";
-                    }
-
-                  }
-                }
-                ,
-
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: TextFormField(
-
-                controller: passwordCtr,
-                autovalidateMode: AutovalidateMode.onUserInteraction,
-                style: TextStyle(color: Colors.black),
-                decoration: InputDecoration(
-                  hintText: "password",
-                  labelText: "password",
+                  hintText: "price",
+                  labelText: "price",
                   labelStyle: TextStyle(fontWeight: FontWeight.bold),
                   counterText: "20",
                   border:
@@ -295,68 +256,85 @@ class _userAdd extends State<userAdd> {
             ),
             Padding(
               padding: const EdgeInsets.all(20.0),
-              child: DropdownButtonFormField<String>(
-                value: selectedGender,
-                onChanged: (newValue) {
-                  setState(() {
-                    selectedGender = newValue;
-                    genderCtr.text = newValue ?? '';
-                  });
-                },
+              child: TextFormField(
+
+                controller: locationCtr,
+                autovalidateMode: AutovalidateMode.onUserInteraction,
+                style: TextStyle(color: Colors.black),
                 decoration: InputDecoration(
-                  labelText: 'Gender',
+                  hintText: "location",
+                  labelText: "location",
                   labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                  border: OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.amber),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  errorBorder: OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.red),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
+                  counterText: "20",
+                  border:
+                  OutlineInputBorder(borderSide:BorderSide(color: Colors.amber),
+                      borderRadius: BorderRadius.circular(20)),
+                  errorBorder:  OutlineInputBorder(borderSide:BorderSide(color: Colors.red),
+                      borderRadius: BorderRadius.circular(20)),
+
+
+
                 ),
-                items: genderList.map((gender) {
-                  return DropdownMenuItem<String>(
-                    value: gender,
-                    child: Text(gender),
-                  );
-                }).toList(),
+
+
+                validator: (val){
+                  if(val == ""){
+                    return "value is null";}
+                  if(val != null){
+                    if(val.length <3){
+                      return "password must be more than 3 chars";
+                    }
+
+                  }
+                }
+                ,
+
               ),
             ),
             Padding(
               padding: const EdgeInsets.all(20.0),
-              child: DropdownButtonFormField<String>(
-                value: selectedPermission,
-                onChanged: (newValue) {
-                  setState(() {
-                    selectedPermission = newValue;
-                    perCtr.text = newValue ?? '';
-                  });
-                },
+              child: TextFormField(
+
+                controller: roomsCtr,
+                keyboardType: TextInputType.number,
+                autovalidateMode: AutovalidateMode.onUserInteraction,
+                style: TextStyle(color: Colors.black),
                 decoration: InputDecoration(
-                  labelText: 'Permission',
+                  hintText: "rooms",
+                  labelText: "rooms",
                   labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                  border: OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.amber),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  errorBorder: OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.red),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
+                  counterText: "20",
+                  border:
+                  OutlineInputBorder(borderSide:BorderSide(color: Colors.amber),
+                      borderRadius: BorderRadius.circular(20)),
+                  errorBorder:  OutlineInputBorder(borderSide:BorderSide(color: Colors.red),
+                      borderRadius: BorderRadius.circular(20)),
+
+
+
                 ),
-                items: permissionList.map((permission) {
-                  return DropdownMenuItem<String>(
-                    value: permission,
-                    child: Text(permission),
-                  );
-                }).toList(),
+
+
+                validator: (val){
+                  if(val == ""){
+                    return "value is null";}
+                  if(val != null){
+                    if(val.length <3){
+                      return "rooms must be more than 3 chars";
+                    }
+
+                  }
+                }
+                ,
+
               ),
             ),
 
 
 
-           if (loading) CircularProgressIndicator() else TextButton(onPressed: ()async{
+
+
+            if (loading) CircularProgressIndicator() else TextButton(onPressed: ()async{
               if(formKey.currentState!.validate()){
                 try{
                   setState(() {
@@ -365,19 +343,20 @@ class _userAdd extends State<userAdd> {
                     iserror=false;
 
                   });
-                var date={
-                "fName":nameCtr.text,
-                  "img":"https://th.bing.com/th/id/R.e2981720d54bd5c7869ed4918473dbf5?rik=3km1AVdxxXLKSA&riu=http%3a%2f%2fvbconversions.com%2fwp-content%2fuploads%2f2018%2f04%2fperson-icon-6.png&ehk=N8n%2bOsRYgQcalmQs9Vv9wEsqtw93GDpSp23eQJOwfTM%3d&risl=&pid=ImgRaw&r=0",
-                "uName":nameCtr.text,
-                "password":passwordCtr.text,
-                "phone":int.parse(phoneCtr.text),
-                "per":perCtr.text,
-                "gender":genderCtr.text,
-                "email":emailCtr.text,
+                var data={
+                "name":nameCtr.text,
+                  "avatar":"https://th.bing.com/th/id/R.ce43f7e8e0571c21e762b8924aad874d?rik=0ATDeAN%2bQ7cyxQ&riu=http%3a%2f%2ftravelji.com%2fwp-content%2fuploads%2fHotel-Tips.jpg&ehk=LzuGeOKqbj3J7q%2f%2fldexRjd2c0yRmq%2b%2fypHlVvmA3dM%3d&risl=&pid=ImgRaw&r=0",
+
+                "descnmae":descnmaeCtr.text,
+                "description":descriptionCtr.text,
+                "price":priceCtr.text,
+                "location":locationCtr.text,
+                "rooms":int.parse(roomsCtr.text),
+
 
 
                 };
-                var addRes=await UserRepository().addd(UserModel.fromJson(date));
+                var addRes=await HotelRepository().addd(HotelModel.fromJson(data));
                 if(addRes!=true){
                 setState(() {
                 loading=false;
@@ -411,6 +390,9 @@ class _userAdd extends State<userAdd> {
                 }
 
               }
+
+
+
 
 
                 , child: Text("send")),
