+++ finalproject/lib/views/dashboard/users/edit_user.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
-import '../models/user_model.dart';
-import '../repository/user_repo.dart';
+
+import '../../../models/user_model.dart';
+import '../../../repository/user_repo.dart';
+
 String? selectedPermission;
 final permissionList = ['admin', 'user'];
