+++ finalproject/lib/views/hotel/Hotels_page.dart
@@ -1,0 +1,162 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_rating_bar/flutter_rating_bar.dart';
+import 'package:project4/views/hotel/Hotal_detail.dart';
+
+import '../../models/hotel_model.dart';
+import '../../repository/hotel_repo.dart';
+import '../reservation/add_reservation.dart';
+
+class HotelColumn extends StatefulWidget {
+  @override
+  State<HotelColumn> createState() => _HotelColumnState();
+}
+
+class _HotelColumnState extends State<HotelColumn> {
+  @override
+  Widget build(BuildContext context) {
+    return RefreshIndicator(
+      onRefresh: () async {
+        setState(() {});
+      },
+      child: SingleChildScrollView(
+        child: Column(
+          children: [HotelCard()],
+        ),
+      ),
+    );
+  }
+}
+
+class HotelCard extends StatefulWidget {
+  @override
+  State<HotelCard> createState() => _HotelCardState();
+}
+
+class _HotelCardState extends State<HotelCard> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: FutureBuilder<List<HotelModel>>(
+        future: HotelRepository().getAll(),
+        builder: (context, snapshot) {
+          if (snapshot.connectionState == ConnectionState.waiting) {
+            return Center(child: CircularProgressIndicator());
+          } else if (snapshot.connectionState == ConnectionState.done) {
+            if (snapshot.hasError)
+              return Center(child: Text("Error ${snapshot.error.toString()}"));
+            else if (snapshot.hasData) {
+              var list = snapshot.data ?? [];
+              return ListView.separated(
+                shrinkWrap: true,
+                physics: NeverScrollableScrollPhysics(),
+                itemBuilder: (context, index) {
+                  return Container(
+                    child: Card(
+                      elevation: 4.0,
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(12.0),
+                      ),
+                      child: InkWell(
+                        onTap: () {
+
+                          HotelRepository().getById(list[index].id.toString());
+
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(builder: (context) => ReservationPage()),
+                          );
+
+
+                        },
+                        child: Container(
+                          width: double.infinity,
+                          padding: EdgeInsets.all(16.0),
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              ClipRRect(
+                                borderRadius: BorderRadius.circular(12.0),
+                                child: list[index].avatar == ''
+                                    ? Container(
+                                  child: Icon(Icons.image),
+                                  width: 50,
+                                  height: 50,
+                                )
+                                    : Image.network(
+                                  list[index].avatar!,
+                                  width: double.infinity,
+                                  height: 150.0,
+                                  fit: BoxFit.cover,
+                                ),
+                              ),
+                              SizedBox(height: 8.0),
+                              Row(
+                                mainAxisAlignment:
+                                MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Text(
+                                    "${list[index].name}",
+                                    style: TextStyle(
+                                      fontSize: 20.0,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                  Text(
+                                    '${list[index].price}',
+                                    style: TextStyle(
+                                      fontSize: 18.0,
+                                      fontWeight: FontWeight.bold,
+                                      color: Colors.black,
+                                    ),
+                                  ),
+                                ],
+                              ),
+                              SizedBox(height: 8.0),
+                              Text(
+                                '${list[index].description}',
+                                style: TextStyle(
+                                  fontSize: 16.0,
+                                  color: Colors.grey[600],
+                                ),
+                              ),
+                              SizedBox(height: 8.0),
+                              RatingBar.builder(
+                               // initialRating: list[index].rating,
+                                minRating: 1,
+                                direction: Axis.horizontal,
+                                allowHalfRating: true,
+                                itemCount: 5,
+                                itemSize: 24,
+                                itemBuilder: (context, _) => Icon(
+                                  Icons.star,
+                                  color: Colors.amber,
+                                ),
+                                onRatingUpdate: (rating) {
+                                  // Handle rating update
+                                },
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                  );
+                },
+                separatorBuilder: (context, index) {
+                  return SizedBox(height: 16.0);
+                },
+                itemCount: list.length,
+              );
+            } else {
+              return Center(child: Text("No data available"));
+            }
+          } else {
+            return Center(child: Text("Error loading data"));
+          }
+        },
+      ),
+    );
+  }
+}
