+++ finalproject/lib/views/offer_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:carousel_slider/carousel_slider.dart';
-import 'reservation_page.dart';
+
+import 'reservation/add_reservation.dart';
 
 class OfferPage extends StatelessWidget {
   @override
